fix(sdk): restore login to be not part of sdk
