+++ core/service/refresh.go
@@ -7,7 +7,6 @@ package service
 import (
 	"dhcli/utils"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -36,7 +35,7 @@ func RefreshHandler() error {
 	}
 
 	if resp.StatusCode != http.StatusOK {
-		return errors.New(fmt.Sprintf("Token server error: %s %s", resp.Status, string(body)))
+		return fmt.Errorf("token server error: %s %s", resp.Status, string(body))
 	}
 
 	var responseJson map[string]interface{}
